backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic
pydantic-settings
httpx
python-jose[cryptography]
passlib[bcrypt]
apscheduler
openai
jinja2
python-multipart
